Harden Link Check (#154)
(1) Added comments about linkcheck exceptions
(2) Added Solidworks to linkcheck exceptions - I'm tired of dealing with 403 errors from that website.
(3) Added 15s timeout to linkcheck to prevent it from hanging forever and not showing what made it hang. Now worse-case it "hangs" on a single element for no more than 15 seconds, and then shows what it failed on.
(4) Updated the user-agent. Maybe not necessary, but it's the recommended browser we're checking links against.